+++ Source/RayLibs/Raytracer/Films/InteractiveFilm.cpp
@@ -22,7 +22,7 @@ Film(i_x_resolution, i_y_resolution), m_x_resolution(i_x_resolution), m_y_resolu
   ASSERT(ip_filter != NULL);
 
   m_crop_window_begin = Point2D_i(0, 0);
-  m_crop_window_end = Point2D_i(m_x_resolution, m_y_resolution);
+  m_crop_window_end = Point2D_i((int)m_x_resolution, (int)m_y_resolution);
 
   if (i_x_resolution==0 || i_y_resolution==0) return;
